+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-google-cloud-platform/compare/0.5.0...HEAD
 
 ### Changed
-
+- Renamed field max_instance_request_concurrency in create_service method. Modified init value.
+- Added vpc_access parameter to RevisionTemplate object
+- Added test for 'chaosgcp.cloudrun.actions.update_service'
 - Updated RevisionTemplate field name max_instance_request_concurrency 
 
 ## [0.5.0][] - 2022-06-21